AzureFunctions/requirements.txt
azure-functions
azure-servicebus
azure-cosmos
azure-monitor-query
requests
python-dotenv
azure-identity
pydantic
